+++ .github/scripts/serialization/test_load.py
@@ -1,0 +1,30 @@
+"""
+Load fixtures generated by an old spikeinterface release with the current code and check
+they reload correctly.
+
+This test is workflow-only: it lives outside src/ so the normal test suite does not
+collect it, and cross_version_serialization.yml invokes it by path after generating the
+fixtures (pointed to by SI_SERIALIZATION_FIXTURES_DIR). Because it sits next to battery.py,
+it imports the shared battery directly.
+"""
+
+import os
+from pathlib import Path
+
+import pytest
+
+from spikeinterface.core import load
+
+from battery import BATTERY, FIXTURE_SUFFIX
+
+FIXTURES_DIR = Path(os.environ.get("SI_SERIALIZATION_FIXTURES_DIR", "serialization_fixtures"))
+
+_CASES = [(entry, fmt) for entry in BATTERY for fmt in entry["formats"]]
+
+
+@pytest.mark.parametrize("entry,fmt", _CASES, ids=[f"{entry['id']}-{fmt}" for entry, fmt in _CASES])
+def test_load_old_serialized_object(entry, fmt):
+    fixture = FIXTURES_DIR / f"{entry['id']}{FIXTURE_SUFFIX[fmt]}"
+    assert fixture.exists(), f"missing fixture {fixture}"
+    obj = load(fixture)
+    entry["check"](obj)
