+++ .github/workflows/cross_version_serialization.yml
@@ -15,6 +15,10 @@ on:
     types: [synchronize, opened, reopened]
     branches:
       - main
+    # Only trigger when a change could affect whether old files still load: the code
+    # under test (core), the serialization harness, or this workflow itself. Note this
+    # watches core only; a break introduced in generation/ or preprocessing/ would not
+    # trigger it.
     paths:
       - 'src/spikeinterface/core/**'
       - '.github/scripts/serialization/**'
@@ -28,6 +32,9 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
+  # Compute which released versions the matrix below tests loading from, exposed as the
+  # job output `list`. Latest patch of each minor at or above the floor on schedule /
+  # dispatch; only the latest minor on a pull request (keeps per-PR runs cheap).
   versions:
     name: Select versions
     runs-on: ubuntu-latest
@@ -42,7 +49,7 @@ jobs:
           GITHUB_EVENT_NAME: ${{ github.event_name }}
           # Support floor: oldest minor to test. Below this, installs fail on the CI
           # Python (numcodecs dependency rot in 0.100/0.101). Raise when 0.102 rots.
-          MIN_VERSION_TO_TEST: "0.102"
+          MIN_VERSION_TO_TEST: "0.102"  # We will change this when making breaking changes (hopefully not too often)
         run: |
           python -m pip install -q packaging
           available_versions=$(python -m pip index versions spikeinterface | sed -n 's/.*Available versions: //p')
@@ -103,4 +110,4 @@ jobs:
       - name: Load fixtures with the branch
         run: |
           uv pip install --system -e . --group test-core
-          pytest src/spikeinterface/core/tests/test_cross_version_serialization.py -v
+          pytest .github/scripts/serialization/test_load.py -v
